Remove taskchampion-lib
This was originally intended to be a general-purpose C API for
TaskChampion, but it's really not great. It's also quite difficult to
use as a crate, rather than including the code within the emebedding
project, as Taskwarrior has done. With this change, TaskChampion is just
a Rust library, used by Taskwarrior but usable by anything.

The integration tests mostly tested the C API, and those are removed.
However, two provided integratoin tests for the taskchampion crate, and
those are moved to `taskchampion/tests`.