Prefer `Operation` instead of `SyncOp` in tests
This is a test-only change: all modified code is `#[cfg(test)]`

Lots of tests used `SyncOp`, including some proptests. One of those is
moved out to `tests/syncing-proptest.rs` and updated to use Replicas
intead of TaskDb's. `op.rs` remains in terms of `SyncOp`, since it defines
that type, but converts to `Operation` for tests.

With their uses in tests removed, two `apply_..` methods can finally be
deleted.