Use `Vec<Operation>` instead of newtype `Operations`
Exposing the full functionality of a Vec here is more useful than
utility functions on a newtype, especially when building a Cxx bridge.

This is a big diff, but aside from the main change in `operation.rs`,
consists almost entirely of changing `ops.add` to `ops.push`, with a few
related test changes.