Add support for sync to AWS
This approach is very similar to that of GCP, and utilizes a similar
conditional-write mechanism to prevent conflicting updates to the
`latest` key.

This updates the package major version to 1.0.0, as it modifies the
`ServerConfig` enum by adding a new variant -- a semver-breaking change.
To avoid further major version bumps from additional variants, this also
adds `#[non-exhaustive]` to the enum.