Move the main crate to the top level of the repo
Once upon a time, there were several crates in a workspace here. Now all
that's left is `xtask`, which is just a command-line utility for
development. So, let's move the main code to the top level of the repo.

This has no behavioral change - it is just moving things about.