Introduce an access mode for SQLite storage (#511)
The storage implementation checks this access mode for all methods, but
also opens SQLite in read-only mode when appropriate, which causes
SQLite to skip creating a WAL log or updating the access time in the
database.

As a concession to ease-of-use with a new task database, the database is
opened in read-write mode and the schema set up if it does not exist,
regardless of the passed `AccessMode`.

This is a breaking change because:

 - The publicly-visible method `SqliteStorage::new` now takes a
   different number of parameters.
 - The public enum `StorageConfig` has been marked #[non_exhaustive].
   Pattern-matching on it outside of its crate must now include a
   wildcard pattern like `_`, or it will fail to compile.
 - `StorageConfig::OnDisk` has an additional field, `access_mode`, which
   has to be included when constructing or matching on this variant.