Improve performance of `taskdb::apply_operations` (#533)
The existing implementation performed a `txn.get_task` and
`txn.set_task` for each update operation. But typically, a sequence of
operations will all apply to the same task, so most of these calls are
redundant.

This introduces a small cache of tasks to which some update operations
have been applied (or which are known to not exist), avoiding the
redundant calls.