Deprecated namespaced UDAs; promote legacy UDAs to be default (#540) (#545)
This commit:
- Adds documentation about why namespaced UDAs existed and how they are
  an old idea
- Deprecates functions for namespaced and legacy UDAs
- Adds functions for handling taskwarrior compatible UDAs instead of relying
  on functions using the `legacy` word

About the deprecation of namespaced UDAs:

When TaskChampion was first implemented there was an idea to only use
namespaced keys in the future. This was before TaskChampion was integrated into
Taskwarrior 3, but it lead to UDAs that are currently being used being referenced
as "legacy" UDAs.

This commit changes that idea: Namespaced UDAs are deprecated, and what was called
legacy will be the way for UDAs going forward.