Modify the working set in-place rather than clearing and rewriting (#557)
This saves a lot of time in the common case of no change to the working
set, as no changes are made. But the next-most-common cases are adding a
new item (just one call to `add_to_working_set`) or removing a
no-longer-pending item (a few calls to `set_working_set_item` for items
that have shifted their position.

This should be faster without any functional change.