feat(taskchampion::composing_json): Init
This module can be used to turn the `Task` structure into a line of
json. This is needed to support with, for example, the TW
hooks.

Unfortunately, having these functions also facilitates getting a `Task`
without going through the Replica. This could be an API foot gun or simply
a compatibility hazard, but at the other hand, these functions are quite obviously
only meant for interoperability and not for _creating_ tasks.