Fix incorrect string quoting in schema (#612)
* Fix incorrect string quoting in schema

The use of `"$.Delete.uuid"` and similar in the operations.uuid column
is invalid. SQLite version 3.45.1 at least allows this invalid syntax,
but newer versions do not.

Unfortunately, fixing this requires a schema update, and requires
dropping and re-creating the `operations_by_uuid` index. For users with
a huge number of operations (millions), this could take some time to
complete (10's of seconds).

* Test with the bundled feature, too