Factor server implementations to use a single HTTP library (#645)
* MSRV bump

* Factor server implementations to use a single HTTP library

This refactors all three remote server implementations (AWS, GCP,
Server) to use reqwest. This reduces the number of dependencies and
simplifies configuration of HTTP parameters like TLS certs. It also
makes the `server-sync` implementation actually async, replacing the
blocking use of ureq.

Mostly, this involves some adaptation to make the cloud SDKs use a
provided HTTP client, but this is relatively straightforward. The Server
implementation sees a more substantial overhaul, and in the process gets
tests to ensure it's behaving reasonably.

A simple binary built against the library was 25MB before and 18MB after
this change.

To Do:
 - Use `cargo deny` to deny undesired libraries / versions