Add server-sync support on WASM
This supports server-sync with few changes by virtue of reqwest's WASM
support.

The `wasm-pack test` machinery does not support starting a test server
reachable from the browser, so there is no practical way to test this
functionality on this platform. Since the implementation differs only
slightly from that for non-web platforms, this is acceptable.